// web/src/api/model/domain.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Cubit Control Plane API
 * Strict REST API for Cubit, an open-source bare-metal PaaS for running Cloudflare Workers and Durable Objects using celld, Traefik, and S3 storage.
 * OpenAPI spec version: 1.0.0
 */

export interface Domain {
  id: string;
  applicationId: string;
  hostname: string;
  pathPrefix: string;
  sslActive: boolean;
  createdAt: string;
}